@@ -6,7 +6,7 @@ homepage: https://github.com/GeekyAnts/flick-video-player
 version: 0.1.1
 
 environment:
-  sdk: ">=2.4.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
