@@ -3,7 +3,7 @@ import 'package:flick_video_player/flick_video_player.dart';
 
 /// Default landscape controls.
 class FlickLandscapeControls extends StatelessWidget {
-  const FlickLandscapeControls({Key key}) : super(key: key);
+  const FlickLandscapeControls({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
