@@ -13,7 +13,7 @@ import 'package:video_player/video_player.dart';
 ///      * Controls.
 class FlickVideoWithControls extends StatefulWidget {
   const FlickVideoWithControls({
-    Key key,
+    Key? key,
     this.controls,
     this.videoFit = BoxFit.cover,
     this.playerLoadingFallback = const Center(
@@ -39,7 +39,7 @@ class FlickVideoWithControls extends StatefulWidget {
   }) : super(key: key);
 
   /// Create custom controls or use any of these [FlickPortraitControls], [FlickLandscapeControls]
-  final Widget controls;
+  final Widget? controls;
 
   /// Conditionally rendered if player is not initialized.
   final Widget playerLoadingFallback;
@@ -80,11 +80,11 @@ class FlickVideoWithControls extends StatefulWidget {
 }
 
 class _FlickVideoWithControlsState extends State<FlickVideoWithControls> {
-  VideoPlayerController _videoPlayerController;
+  VideoPlayerController? _videoPlayerController;
 
   @override
   void didChangeDependencies() {
-    VideoPlayerController newController =
+    VideoPlayerController? newController =
         Provider.of<FlickVideoManager>(context).videoPlayerController;
     if ((widget.willVideoPlayerControllerChange &&
             _videoPlayerController != newController) ||
