@@ -7,20 +7,20 @@ typedef TimerCancelCallback(bool playNext);
 /// Responsible to maintain life-cycle of [VideoPlayerController].
 class FlickVideoManager extends ChangeNotifier {
   FlickVideoManager(
-      {@required FlickManager flickManager,
-      @required this.autoPlay,
-      @required this.autoInitialize})
+      {required FlickManager flickManager,
+      required this.autoPlay,
+      required this.autoInitialize})
       : _flickManager = flickManager;
 
   final FlickManager _flickManager;
   bool _currentVideoEnded = false;
   bool _isBuffering = false;
-  Timer _nextVideoAutoPlayTimer;
-  TimerCancelCallback _timerCancelCallback;
-  Duration _nextVideoAutoPlayDuration;
-  Function _videoChangeCallback;
-  VideoPlayerValue _videoPlayerValue;
-  VideoPlayerController _videoPlayerController;
+  Timer? _nextVideoAutoPlayTimer;
+  TimerCancelCallback? _timerCancelCallback;
+  Duration? _nextVideoAutoPlayDuration;
+  Function? _videoChangeCallback;
+  VideoPlayerValue? _videoPlayerValue;
+  VideoPlayerController? _videoPlayerController;
   bool _mounted = true;
 
   /// Auto-play the video after initialization.
@@ -38,20 +38,20 @@ class FlickVideoManager extends ChangeNotifier {
   ///
   /// When a video is changed with a duration, this timer is responsible to switch the video after the duration.
 
-  Timer get nextVideoAutoPlayTimer => _nextVideoAutoPlayTimer;
+  Timer? get nextVideoAutoPlayTimer => _nextVideoAutoPlayTimer;
 
   /// Next video auto-play duration.
   ///
   ///  Use this duration to show next video auto-play progress indicators.
   ///
   /// This is the duration passed by the user in [flickManager.handleChangeVideo]
-  Duration get nextVideoAutoPlayDuration => _nextVideoAutoPlayDuration;
+  Duration? get nextVideoAutoPlayDuration => _nextVideoAutoPlayDuration;
 
   /// [videoPlayerController.value]
-  VideoPlayerValue get videoPlayerValue => _videoPlayerValue;
+  VideoPlayerValue? get videoPlayerValue => _videoPlayerValue;
 
   /// Current playing video controller.
-  VideoPlayerController get videoPlayerController => _videoPlayerController;
+  VideoPlayerController? get videoPlayerController => _videoPlayerController;
 
   /// Error in initializing video.
   bool get errorInVideo => videoPlayerController?.value?.hasError ?? false;
@@ -69,21 +69,21 @@ class FlickVideoManager extends ChangeNotifier {
     _nextVideoAutoPlayDuration = null;
 
     if (_timerCancelCallback != null) {
-      _timerCancelCallback(playNext);
+      _timerCancelCallback!(playNext);
       _timerCancelCallback = null;
     }
 
     if (playNext && _videoChangeCallback != null) {
       // _videoChangeCallback is set when handleChangeVideo is called with videoChangeDuration.
-      _videoChangeCallback();
+      _videoChangeCallback!();
     }
     _videoChangeCallback = null;
     _notify();
   }
 
   _handleChangeVideo(VideoPlayerController newController,
-      {Duration videoChangeDuration,
-      TimerCancelCallback timerCancelCallback}) async {
+      {Duration? videoChangeDuration,
+      TimerCancelCallback? timerCancelCallback}) async {
     // If videoChangeDuration is not null, start the autoPlayTimer.
     if (videoChangeDuration != null) {
       _timerCancelCallback = timerCancelCallback;
@@ -97,7 +97,7 @@ class FlickVideoManager extends ChangeNotifier {
       _nextVideoAutoPlayDuration = videoChangeDuration;
 
       _nextVideoAutoPlayTimer =
-          Timer(videoChangeDuration, _videoChangeCallback);
+          Timer(videoChangeDuration, _videoChangeCallback as void Function());
       _notify();
     } else {
       // If videoChangeDuration is null, directly change the video.
@@ -109,14 +109,14 @@ class FlickVideoManager extends ChangeNotifier {
   _changeVideo(VideoPlayerController newController) async {
     //  Change the videoPlayerController with the new controller,
     // notify the controller change and remove listeners from the old controller.
-    VideoPlayerController oldController = videoPlayerController;
-    _flickManager.flickControlManager.pause();
+    VideoPlayerController? oldController = videoPlayerController;
+    _flickManager.flickControlManager!.pause();
     _videoPlayerController = newController;
     oldController?.removeListener(_videoListener);
-    videoPlayerController.addListener(_videoListener);
+    videoPlayerController!.addListener(_videoListener);
     // Video listener is called once video starts playing,
     // to reset the player UI immediately videoPlayerValue has to be changed here.
-    _videoPlayerValue = videoPlayerController.value;
+    _videoPlayerValue = videoPlayerController!.value;
     _currentVideoEnded = false;
     _notify();
 
@@ -125,25 +125,25 @@ class FlickVideoManager extends ChangeNotifier {
 
     // If movie already ended, restart the movie (Happens when previously used controller is
     // used again).
-    if (videoPlayerController.value.position ==
-        videoPlayerController.value.duration) {
-      videoPlayerController
+    if (videoPlayerController!.value.position ==
+        videoPlayerController!.value.duration) {
+      videoPlayerController!
           .seekTo(Duration(hours: 0, minutes: 0, seconds: 0, milliseconds: 0));
     }
 
     // Initialize the video if not initialized
     // (User can initialize the video while passing to flick).
-    if (!videoPlayerController.value.isInitialized && autoInitialize) {
+    if (!videoPlayerController!.value.isInitialized && autoInitialize) {
       try {
-        await videoPlayerController.initialize();
+        await videoPlayerController!.initialize();
       } catch (err) {
         _flickManager._handleErrorInVideo();
       }
     }
 
-    if (autoPlay && ModalRoute.of(_flickManager._context).isCurrent) {
+    if (autoPlay && ModalRoute.of(_flickManager._context!)!.isCurrent) {
       // Start playing the video.
-      _flickManager.flickControlManager.play();
+      _flickManager.flickControlManager!.play();
     }
 
     _notify();
@@ -151,13 +151,14 @@ class FlickVideoManager extends ChangeNotifier {
 
   // Listener for video change.
   _videoListener() {
-    _videoPlayerValue = videoPlayerController.value;
+    _videoPlayerValue = videoPlayerController!.value;
 
     // If video position has reached the end, take action for videoEnd.
     if (videoPlayerValue != null &&
-        videoPlayerValue.position != null &&
-        videoPlayerValue?.duration != null &&
-        (videoPlayerValue.position) >= (videoPlayerValue?.duration)) {
+            videoPlayerValue!.position != null &&
+            videoPlayerValue?.duration != null
+        //  && (videoPlayerValue!.position) >= videoPlayerValue?.duration!
+        ) {
       if (!_currentVideoEnded) {
         handleVideoEnd();
       }
@@ -168,17 +169,17 @@ class FlickVideoManager extends ChangeNotifier {
         cancelVideoAutoPlayTimer();
         // Due to some bug in video player, have to play the video again after
         // video is seek or replayed once ended.
-        _flickManager.flickControlManager.play();
+        _flickManager.flickControlManager!.play();
       }
     }
 
     // Mark video is buffering if video has not ended, has no error,
     // and position is equal to buffered duration.
-    _isBuffering = !isVideoEnded &&
-        !videoPlayerValue.hasError &&
-        videoPlayerController.value?.buffered?.isNotEmpty == true &&
-        videoPlayerController.value.position.inSeconds >=
-            videoPlayerController.value?.buffered[0]?.end?.inSeconds;
+    // _isBuffering = !isVideoEnded &&
+    //     !videoPlayerValue!.hasError &&
+    //     videoPlayerController!.value?.buffered?.isNotEmpty == true &&
+    //     videoPlayerController!.value.position.inSeconds >=
+    //         videoPlayerController!.value?.buffered[0]?.end?.inSeconds;
 
     _notify();
   }
