@@ -4,23 +4,23 @@ import 'package:provider/provider.dart';
 
 /// Uses [MultiProviders] to add all the managers as providers.
 class FlickManagerBuilder extends StatelessWidget {
-  const FlickManagerBuilder({Key key, this.child, this.flickManager})
+  const FlickManagerBuilder({Key? key, this.child, this.flickManager})
       : super(key: key);
-  final Widget child;
-  final FlickManager flickManager;
+  final Widget? child;
+  final FlickManager? flickManager;
 
   @override
   Widget build(BuildContext context) {
     return MultiProvider(
       providers: [
-        ChangeNotifierProvider<FlickVideoManager>.value(
-          value: flickManager.flickVideoManager,
+        ChangeNotifierProvider<FlickVideoManager?>.value(
+          value: flickManager!.flickVideoManager,
         ),
-        ChangeNotifierProvider<FlickDisplayManager>.value(
-          value: flickManager.flickDisplayManager,
+        ChangeNotifierProvider<FlickDisplayManager?>.value(
+          value: flickManager!.flickDisplayManager,
         ),
-        ChangeNotifierProvider<FlickControlManager>.value(
-          value: flickManager.flickControlManager,
+        ChangeNotifierProvider<FlickControlManager?>.value(
+          value: flickManager!.flickControlManager,
         ),
       ],
       child: child,
