@@ -7,7 +7,7 @@ import 'package:provider/provider.dart';
 /// Renders two GestureDetector inside a row, the first detector is responsible to seekBackward and the second detector is responsible to seekForward.
 class FlickSeekVideoAction extends StatelessWidget {
   const FlickSeekVideoAction({
-    Key key,
+    Key? key,
     this.child,
     this.forwardSeekIcon = const Icon(Icons.fast_forward),
     this.backwardSeekIcon = const Icon(Icons.fast_rewind),
@@ -17,7 +17,7 @@ class FlickSeekVideoAction extends StatelessWidget {
   }) : super(key: key);
 
   /// Widget to be stacked above this action.
-  final Widget child;
+  final Widget? child;
 
   /// Widget to be shown when user forwardSeek the video.
   ///
@@ -35,7 +35,7 @@ class FlickSeekVideoAction extends StatelessWidget {
   /// ``` dart
   ///    controlManager.seekForward(Duration(seconds: 10));
   /// ```
-  final Function seekForward;
+  final Function? seekForward;
 
   /// Function called onTap of [backwardSeekIcon].
   ///
@@ -43,7 +43,7 @@ class FlickSeekVideoAction extends StatelessWidget {
   /// ``` dart
   ///     controlManager.seekBackward(Duration(seconds: 10));
   /// ```
-  final Function seekBackward;
+  final Function? seekBackward;
 
   /// Duration by which video will be seek.
   final Duration duration;
@@ -58,73 +58,68 @@ class FlickSeekVideoAction extends StatelessWidget {
     bool showForwardSeek = displayManager.showForwardSeek;
     bool showBackwardSeek = displayManager.showBackwardSeek;
 
-    return Stack(
+    return Row(
+      crossAxisAlignment: CrossAxisAlignment.stretch,
       children: <Widget>[
-        Row(
-          crossAxisAlignment: CrossAxisAlignment.stretch,
-          children: <Widget>[
-            Expanded(
-              child: GestureDetector(
-                behavior: HitTestBehavior.opaque,
-                onDoubleTap: () {
-                  if (seekBackward != null) {
-                    seekBackward();
-                  } else {
-                    controlManager.seekBackward(duration);
-                  }
-                },
-                child: Align(
-                  alignment: Alignment.center,
-                  child: AnimatedCrossFade(
-                    duration: Duration(milliseconds: 100),
-                    firstChild: IconTheme(
-                      data: IconThemeData(color: Colors.transparent),
-                      child: backwardSeekIcon,
-                    ),
-                    crossFadeState: showBackwardSeek
-                        ? CrossFadeState.showSecond
-                        : CrossFadeState.showFirst,
-                    secondChild: IconTheme(
-                      data: IconThemeData(color: Colors.white),
-                      child: backwardSeekIcon,
-                    ),
-                  ),
+        Expanded(
+          child: GestureDetector(
+            behavior: HitTestBehavior.opaque,
+            onDoubleTap: () {
+              if (seekBackward != null) {
+                seekBackward!();
+              } else {
+                controlManager.seekBackward(duration);
+              }
+            },
+            child: Align(
+              alignment: Alignment.center,
+              child: AnimatedCrossFade(
+                duration: Duration(milliseconds: 100),
+                firstChild: IconTheme(
+                  data: IconThemeData(color: Colors.transparent),
+                  child: backwardSeekIcon,
+                ),
+                crossFadeState: showBackwardSeek
+                    ? CrossFadeState.showSecond
+                    : CrossFadeState.showFirst,
+                secondChild: IconTheme(
+                  data: IconThemeData(color: Colors.white),
+                  child: backwardSeekIcon,
                 ),
               ),
             ),
-            Expanded(
-              child: GestureDetector(
-                behavior: HitTestBehavior.opaque,
-                onDoubleTap: () {
-                  if (seekForward != null) {
-                    seekForward();
-                  } else {
-                    controlManager.seekForward(duration);
-                  }
-                },
-                child: Align(
-                  alignment: Alignment.center,
-                  child: AnimatedCrossFade(
-                    duration: Duration(milliseconds: 100),
-                    firstChild: IconTheme(
-                        data: IconThemeData(
-                          color: Colors.transparent,
-                        ),
-                        child: forwardSeekIcon),
-                    crossFadeState: showForwardSeek
-                        ? CrossFadeState.showSecond
-                        : CrossFadeState.showFirst,
-                    secondChild: IconTheme(
-                      data: IconThemeData(color: Colors.white),
-                      child: forwardSeekIcon,
+          ),
+        ),
+        Expanded(
+          child: GestureDetector(
+            behavior: HitTestBehavior.opaque,
+            onDoubleTap: () {
+              if (seekForward != null) {
+                seekForward!();
+              } else {
+                controlManager.seekForward(duration);
+              }
+            },
+            child: Align(
+              alignment: Alignment.center,
+              child: AnimatedCrossFade(
+                duration: Duration(milliseconds: 100),
+                firstChild: IconTheme(
+                    data: IconThemeData(
+                      color: Colors.transparent,
                     ),
-                  ),
+                    child: forwardSeekIcon),
+                crossFadeState: showForwardSeek
+                    ? CrossFadeState.showSecond
+                    : CrossFadeState.showFirst,
+                secondChild: IconTheme(
+                  data: IconThemeData(color: Colors.white),
+                  child: forwardSeekIcon,
                 ),
               ),
-            )
-          ],
-        ),
-        if (child != null) child,
+            ),
+          ),
+        )
       ],
     );
   }
