@@ -5,12 +5,12 @@ import 'package:provider/provider.dart';
 /// GestureDetector that calls [flickDisplayManager.togglePlay] onTap of opaque area/child.
 class FlickTogglePlayAction extends StatelessWidget {
   const FlickTogglePlayAction(
-      {Key key,
+      {Key? key,
       this.child,
       this.behavior = HitTestBehavior.opaque,
       this.togglePlay})
       : super(key: key);
-  final Widget child;
+  final Widget? child;
   final HitTestBehavior behavior;
 
   /// Function called onTap of the opaque area/child.
@@ -19,7 +19,7 @@ class FlickTogglePlayAction extends StatelessWidget {
   /// ``` dart
   ///    displayManager.togglePlay();
   /// ```
-  final Function togglePlay;
+  final Function? togglePlay;
 
   @override
   Widget build(BuildContext context) {
@@ -32,7 +32,7 @@ class FlickTogglePlayAction extends StatelessWidget {
         behavior: behavior,
         onTap: () {
           if (togglePlay != null) {
-            togglePlay();
+            togglePlay!();
           } else {
             controlManager.togglePlay();
           }
